@@ -1,11 +1,13 @@
 """Step arguments tests."""
 
+from __future__ import annotations
+
 import textwrap
 
 
-def test_every_step_takes_param_with_the_same_name(testdir):
+def test_every_step_takes_param_with_the_same_name(pytester):
     """Test every step takes param with the same name."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         arguments=textwrap.dedent(
             """\
@@ -15,13 +17,13 @@ def test_every_step_takes_param_with_the_same_name(testdir):
                     When I pay 2 Euro
                     And I pay 1 Euro
                     Then I should have 0 Euro
-                    And I should have 999999 Euro # In my dream...
+                    And I should have 999999 Euro
 
             """
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import pytest
@@ -37,29 +39,29 @@ def test_every_step_takes_param_with_the_same_name(testdir):
 
 
         @given(parsers.cfparse("I have {euro:d} Euro"))
-        def i_have(euro, values):
+        def _(euro, values):
             assert euro == values.pop(0)
 
 
         @when(parsers.cfparse("I pay {euro:d} Euro"))
-        def i_pay(euro, values, request):
+        def _(euro, values, request):
             assert euro == values.pop(0)
 
 
         @then(parsers.cfparse("I should have {euro:d} Euro"))
-        def i_should_have(euro, values):
+        def _(euro, values):
             assert euro == values.pop(0)
 
         """
         )
     )
-    result = testdir.runpytest()
+    result = pytester.runpytest()
     result.assert_outcomes(passed=1)
 
 
-def test_argument_in_when(testdir):
+def test_argument_in_when(pytester):
     """Test step arguments in when steps."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         arguments=textwrap.dedent(
             """\
@@ -72,7 +74,7 @@ def test_argument_in_when(testdir):
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import pytest
@@ -89,21 +91,21 @@ def test_argument_in_when(testdir):
 
 
         @given(parsers.cfparse("I have an argument {arg:Number}", extra_types=dict(Number=int)))
-        def argument(arguments, arg):
+        def _(arguments, arg):
             arguments["arg"] = arg
 
 
         @when(parsers.cfparse("I get argument {arg:d}"))
-        def get_argument(arguments, arg):
+        def _(arguments, arg):
             arguments["arg"] = arg
 
 
         @then(parsers.cfparse("My argument should be {arg:d}"))
-        def assert_that_my_argument_is_arg(arguments, arg):
+        def _(arguments, arg):
             assert arguments["arg"] == arg
 
         """
         )
     )
-    result = testdir.runpytest()
+    result = pytester.runpytest()
     result.assert_outcomes(passed=1)
