@@ -1,22 +1,16 @@
-import pytest
+from __future__ import annotations
 
-from tests.utils import PYTEST_6
+import pytest
 
 pytest_plugins = "pytester"
 
 
 def pytest_generate_tests(metafunc):
     if "pytest_params" in metafunc.fixturenames:
-        if PYTEST_6:
-            parametrizations = [
-                pytest.param([], id="no-import-mode"),
-                pytest.param(["--import-mode=prepend"], id="--import-mode=prepend"),
-                pytest.param(["--import-mode=append"], id="--import-mode=append"),
-                pytest.param(["--import-mode=importlib"], id="--import-mode=importlib"),
-            ]
-        else:
-            parametrizations = [[]]
-        metafunc.parametrize(
-            "pytest_params",
-            parametrizations,
-        )
+        parametrizations = [
+            pytest.param([], id="no-import-mode"),
+            pytest.param(["--import-mode=prepend"], id="--import-mode=prepend"),
+            pytest.param(["--import-mode=append"], id="--import-mode=append"),
+            pytest.param(["--import-mode=importlib"], id="--import-mode=importlib"),
+        ]
+        metafunc.parametrize("pytest_params", parametrizations)
