@@ -1,13 +1,20 @@
 """Test cucumber json output."""
+
+from __future__ import annotations
+
 import json
 import os.path
 import textwrap
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    from _pytest.pytester import Pytester, RunResult
 
 
-def runandparse(testdir, *args):
+def runandparse(pytester: Pytester, *args: Any) -> tuple[RunResult, list[dict[str, Any]]]:
     """Run tests in testdir and parse json output."""
-    resultpath = testdir.tmpdir.join("cucumber.json")
-    result = testdir.runpytest(f"--cucumberjson={resultpath}", "-s", *args)
+    resultpath = pytester.path.joinpath("cucumber.json")
+    result = pytester.runpytest(f"--cucumberjson={resultpath}", "-s", *args)
     with resultpath.open() as f:
         jsonobject = json.load(f)
     return result, jsonobject
@@ -16,16 +23,16 @@ def runandparse(testdir, *args):
 class OfType:
     """Helper object to help compare object type to initialization type"""
 
-    def __init__(self, type=None):
+    def __init__(self, type: type | None = None) -> None:
         self.type = type
 
-    def __eq__(self, other):
+    def __eq__(self, other: object) -> bool:
         return isinstance(other, self.type) if self.type else True
 
 
-def test_step_trace(testdir):
+def test_step_trace(pytester):
     """Test step trace."""
-    testdir.makefile(
+    pytester.makefile(
         ".ini",
         pytest=textwrap.dedent(
             """
@@ -38,15 +45,18 @@ def test_step_trace(testdir):
     """
         ),
     )
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         test=textwrap.dedent(
             """
     @feature-tag
     Feature: One passing scenario, one failing scenario
+    This is a feature description
 
         @scenario-passing-tag
         Scenario: Passing
+            This is a scenario description
+
             Given a passing step
             And some other passing step
 
@@ -56,7 +66,7 @@ def test_step_trace(testdir):
             And a failing step
 
         @scenario-outline-passing-tag
-        Scenario: Passing outline
+        Scenario Outline: Passing outline
             Given type <type> and value <value>
 
             Examples: example1
@@ -67,26 +77,26 @@ def test_step_trace(testdir):
     """
         ),
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """
         import pytest
-        from pytest_bdd import given, when, scenario
+        from pytest_bdd import given, when, scenario, parsers
 
         @given('a passing step')
-        def a_passing_step():
+        def _():
             return 'pass'
 
         @given('some other passing step')
-        def some_other_passing_step():
+        def _():
             return 'pass'
 
         @given('a failing step')
-        def a_failing_step():
+        def _():
             raise Exception('Error')
 
-        @given('type <type> and value <value>')
-        def type_type_and_value_value():
+        @given(parsers.parse('type {type} and value {value}'))
+        def _():
             return 'pass'
 
         @scenario('test.feature', 'Passing')
@@ -103,112 +113,114 @@ def test_step_trace(testdir):
     """
         )
     )
-    result, jsonobject = runandparse(testdir)
+    result, jsonobject = runandparse(pytester)
+    result.assert_outcomes(passed=4, failed=1)
+
     assert result.ret
     expected = [
         {
-            "description": "",
+            "description": "This is a feature description",
             "elements": [
                 {
-                    "description": "",
+                    "description": "This is a scenario description",
                     "id": "test_passing",
                     "keyword": "Scenario",
-                    "line": 5,
+                    "line": 6,
                     "name": "Passing",
                     "steps": [
                         {
                             "keyword": "Given",
-                            "line": 6,
+                            "line": 9,
                             "match": {"location": ""},
                             "name": "a passing step",
                             "result": {"status": "passed", "duration": OfType(int)},
                         },
                         {
                             "keyword": "And",
-                            "line": 7,
+                            "line": 10,
                             "match": {"location": ""},
                             "name": "some other passing step",
                             "result": {"status": "passed", "duration": OfType(int)},
                         },
                     ],
-                    "tags": [{"name": "scenario-passing-tag", "line": 4}],
+                    "tags": [{"name": "scenario-passing-tag", "line": 5}],
                     "type": "scenario",
                 },
                 {
                     "description": "",
                     "id": "test_failing",
                     "keyword": "Scenario",
-                    "line": 10,
+                    "line": 13,
                     "name": "Failing",
                     "steps": [
                         {
                             "keyword": "Given",
-                            "line": 11,
+                            "line": 14,
                             "match": {"location": ""},
                             "name": "a passing step",
                             "result": {"status": "passed", "duration": OfType(int)},
                         },
                         {
                             "keyword": "And",
-                            "line": 12,
+                            "line": 15,
                             "match": {"location": ""},
                             "name": "a failing step",
                             "result": {"error_message": OfType(str), "status": "failed", "duration": OfType(int)},
                         },
                     ],
-                    "tags": [{"name": "scenario-failing-tag", "line": 9}],
+                    "tags": [{"name": "scenario-failing-tag", "line": 12}],
                     "type": "scenario",
                 },
                 {
                     "description": "",
-                    "keyword": "Scenario",
-                    "tags": [{"line": 14, "name": "scenario-outline-passing-tag"}],
+                    "keyword": "Scenario Outline",
+                    "tags": [{"line": 17, "name": "scenario-outline-passing-tag"}],
                     "steps": [
                         {
-                            "line": 16,
+                            "line": 19,
                             "match": {"location": ""},
                             "result": {"status": "passed", "duration": OfType(int)},
                             "keyword": "Given",
-                            "name": "type <type> and value <value>",
+                            "name": "type str and value hello",
                         }
                     ],
-                    "line": 15,
+                    "line": 18,
                     "type": "scenario",
                     "id": "test_passing_outline[str-hello]",
                     "name": "Passing outline",
                 },
                 {
                     "description": "",
-                    "keyword": "Scenario",
-                    "tags": [{"line": 14, "name": "scenario-outline-passing-tag"}],
+                    "keyword": "Scenario Outline",
+                    "tags": [{"line": 17, "name": "scenario-outline-passing-tag"}],
                     "steps": [
                         {
-                            "line": 16,
+                            "line": 19,
                             "match": {"location": ""},
                             "result": {"status": "passed", "duration": OfType(int)},
                             "keyword": "Given",
-                            "name": "type <type> and value <value>",
+                            "name": "type int and value 42",
                         }
                     ],
-                    "line": 15,
+                    "line": 18,
                     "type": "scenario",
                     "id": "test_passing_outline[int-42]",
                     "name": "Passing outline",
                 },
                 {
                     "description": "",
-                    "keyword": "Scenario",
-                    "tags": [{"line": 14, "name": "scenario-outline-passing-tag"}],
+                    "keyword": "Scenario Outline",
+                    "tags": [{"line": 17, "name": "scenario-outline-passing-tag"}],
                     "steps": [
                         {
-                            "line": 16,
+                            "line": 19,
                             "match": {"location": ""},
                             "result": {"status": "passed", "duration": OfType(int)},
                             "keyword": "Given",
-                            "name": "type <type> and value <value>",
+                            "name": "type float and value 1.0",
                         }
                     ],
-                    "line": 15,
+                    "line": 18,
                     "type": "scenario",
                     "id": "test_passing_outline[float-1.0]",
                     "name": "Passing outline",
@@ -216,110 +228,12 @@ def test_step_trace(testdir):
             ],
             "id": os.path.join("test_step_trace0", "test.feature"),
             "keyword": "Feature",
+            "language": "en",
             "line": 2,
             "name": "One passing scenario, one failing scenario",
             "tags": [{"name": "feature-tag", "line": 1}],
-            "uri": os.path.join(testdir.tmpdir.basename, "test.feature"),
+            "uri": os.path.join(pytester.path.name, "test.feature"),
         }
     ]
 
     assert jsonobject == expected
-
-
-def test_step_trace_with_expand_option(testdir):
-    """Test step trace."""
-    testdir.makefile(
-        ".ini",
-        pytest=textwrap.dedent(
-            """
-    [pytest]
-    markers =
-        feature-tag
-        scenario-outline-passing-tag
-    """
-        ),
-    )
-    testdir.makefile(
-        ".feature",
-        test=textwrap.dedent(
-            """
-    @feature-tag
-    Feature: One scenario outline, expanded to multiple scenarios
-
-        @scenario-outline-passing-tag
-        Scenario: Passing outline
-            Given type <type> and value <value>
-
-            Examples: example1
-            | type    | value  |
-            | str     | hello  |
-            | int     | 42     |
-            | float   | 1.0    |
-    """
-        ),
-    )
-    testdir.makepyfile(
-        textwrap.dedent(
-            """
-        import pytest
-        from pytest_bdd import given, scenario
-
-        @given('type <type> and value <value>')
-        def type_type_and_value_value():
-            return 'pass'
-
-        @scenario('test.feature', 'Passing outline')
-        def test_passing_outline():
-            pass
-    """
-        )
-    )
-    result, jsonobject = runandparse(testdir, "--cucumber-json-expanded")
-    result.assert_outcomes(passed=3)
-
-    assert jsonobject[0]["elements"][0]["steps"][0]["name"] == "type str and value hello"
-    assert jsonobject[0]["elements"][1]["steps"][0]["name"] == "type int and value 42"
-    assert jsonobject[0]["elements"][2]["steps"][0]["name"] == "type float and value 1.0"
-
-
-def test_converters_dict_with_expand_option(testdir):
-    """Test that `--cucumber-json-expanded` works correctly when using `example_converters`."""
-    testdir.makefile(
-        ".feature",
-        test=textwrap.dedent(
-            """
-    Feature: Expanded option with example converters
-        Scenario: Passing outline
-            Given there is an intvalue <intvalue> and stringvalue <stringvalue> and floatvalue <floatvalue>
-
-            Examples: example1
-            | intvalue | stringvalue | floatvalue |
-            | 1        | hello       | 1.0        |
-    """
-        ),
-    )
-    testdir.makepyfile(
-        textwrap.dedent(
-            """
-        import pytest
-        from pytest_bdd import given, scenario
-
-        @given('there is an intvalue <intvalue> and stringvalue <stringvalue> and floatvalue <floatvalue>')
-        def type_type_and_value_value():
-            pass
-
-        @scenario(
-            'test.feature',
-            'Passing outline',
-            example_converters={"intvalue":int, "stringvalue":str, "floatvalue":float},
-        )
-        def test_passing_outline():
-            pass
-    """
-        )
-    )
-    result, jsonobject = runandparse(testdir, "--cucumber-json-expanded")
-    assert result.ret == 0
-
-    expanded_step_name = jsonobject[0]["elements"][0]["steps"][0]["name"]
-    assert expanded_step_name == "there is an intvalue 1 and stringvalue hello and floatvalue 1.0"
