@@ -1,11 +1,13 @@
 """Test descriptions."""
 
+from __future__ import annotations
+
 import textwrap
 
 
-def test_description(testdir):
+def test_description(pytester):
     """Test description for the feature."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         description=textwrap.dedent(
             """\
@@ -19,16 +21,21 @@ def test_description(testdir):
             Some description goes here.
 
             Scenario: Description
+                Also, the scenario can have a description.
+
+                It goes here between the scenario name
+                and the first step.
                 Given I have a bar
         """
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
-            """\
+            r'''
         import textwrap
         from pytest_bdd import given, scenario
+        from pytest_bdd.scenario import scenario_wrapper_template_registry
 
         @scenario("description.feature", "Description")
         def test_description():
@@ -36,22 +43,23 @@ def test_description(testdir):
 
 
         @given("I have a bar")
-        def bar():
+        def _():
             return "bar"
 
-        def test_scenario_description():
-            assert test_description.__scenario__.feature.description == textwrap.dedent(
-                \"\"\"\\
-                In order to achieve something
-                I want something
-                Because it will be cool
-
-
-                Some description goes here.\"\"\"
+        def test_feature_description():
+            scenario = scenario_wrapper_template_registry[test_description]
+            assert scenario.feature.description == textwrap.dedent(
+                "In order to achieve something\nI want something\nBecause it will be cool\n\n\nSome description goes here."
             )
-        """
+
+        def test_scenario_description():
+            scenario = scenario_wrapper_template_registry[test_description]
+            assert scenario.description == textwrap.dedent(
+                "Also, the scenario can have a description.\n\nIt goes here between the scenario name\nand the first step."""
+            )
+        '''
         )
     )
 
-    result = testdir.runpytest()
-    result.assert_outcomes(passed=2)
+    result = pytester.runpytest()
+    result.assert_outcomes(passed=3)
