@@ -1,42 +1,46 @@
 """Test scenarios shortcut."""
+
+from __future__ import annotations
+
 import textwrap
 
-from tests.utils import assert_outcomes
 
-
-def test_scenarios(testdir, pytest_params):
+def test_scenarios(pytester, pytest_params):
     """Test scenarios shortcut (used together with @scenario for individual test override)."""
-    testdir.makeini(
+    pytester.makeini(
         """
             [pytest]
             console_output_style=classic
         """
     )
-    testdir.makeconftest(
+    pytester.makeconftest(
         """
         import pytest
         from pytest_bdd import given
 
         @given('I have a bar')
-        def i_have_bar():
+        def _():
             print('bar!')
             return 'bar'
     """
     )
-    features = testdir.mkdir("features")
-    features.join("test.feature").write_text(
+    features = pytester.mkdir("features")
+    features.joinpath("test.feature").write_text(
         textwrap.dedent(
             """
+Feature: Test scenarios
     Scenario: Test scenario
         Given I have a bar
     """
         ),
         "utf-8",
-        ensure=True,
     )
-    features.join("subfolder", "test.feature").write_text(
+    subfolder = features.joinpath("subfolder")
+    subfolder.mkdir()
+    subfolder.joinpath("test.feature").write_text(
         textwrap.dedent(
             """
+Feature: Test scenarios
     Scenario: Test subfolder scenario
         Given I have a bar
 
@@ -51,9 +55,8 @@ def test_scenarios(testdir, pytest_params):
     """
         ),
         "utf-8",
-        ensure=True,
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         """
         import pytest
         from pytest_bdd import scenarios, scenario
@@ -65,8 +68,8 @@ def test_scenarios(testdir, pytest_params):
         scenarios('features')
     """
     )
-    result = testdir.runpytest_subprocess("-v", "-s", *pytest_params)
-    assert_outcomes(result, passed=4, failed=1)
+    result = pytester.runpytest_subprocess("-v", "-s", *pytest_params)
+    result.assert_outcomes(passed=4, failed=1)
     result.stdout.fnmatch_lines(["*collected 5 items"])
     result.stdout.fnmatch_lines(["*test_test_subfolder_scenario *bar!", "PASSED"])
     result.stdout.fnmatch_lines(["*test_test_scenario *bar!", "PASSED"])
@@ -75,9 +78,9 @@ def test_scenarios(testdir, pytest_params):
     result.stdout.fnmatch_lines(["*test_test_scenario_1 *bar!", "PASSED"])
 
 
-def test_scenarios_none_found(testdir, pytest_params):
+def test_scenarios_none_found(pytester, pytest_params):
     """Test scenarios shortcut when no scenarios found."""
-    testpath = testdir.makepyfile(
+    testpath = pytester.makepyfile(
         """
         import pytest
         from pytest_bdd import scenarios
@@ -85,6 +88,6 @@ def test_scenarios_none_found(testdir, pytest_params):
         scenarios('.')
     """
     )
-    result = testdir.runpytest_subprocess(testpath, *pytest_params)
-    assert_outcomes(result, errors=1)
+    result = pytester.runpytest_subprocess(testpath, *pytest_params)
+    result.assert_outcomes(errors=1)
     result.stdout.fnmatch_lines(["*NoScenariosFound*"])
