@@ -1,15 +1,13 @@
 """Test tags."""
+
+from __future__ import annotations
+
 import textwrap
 
-import pkg_resources
-import pytest
 
-from pytest_bdd.parser import get_tags
-
-
-def test_tags_selector(testdir):
+def test_tags_selector(pytester):
     """Test tests selection by tags."""
-    testdir.makefile(
+    pytester.makefile(
         ".ini",
         pytest=textwrap.dedent(
             """
@@ -24,53 +22,59 @@ def test_tags_selector(testdir):
     """
         ),
     )
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         test="""
     @feature_tag_1 @feature_tag_2
     Feature: Tags
 
-    @scenario_tag_01 @scenario_tag_02
-    Scenario: Tags
-        Given I have a bar
+        @scenario_tag_01 @scenario_tag_02
+        Scenario: Tags
+            Given I have a bar
 
-    @scenario_tag_10 @scenario_tag_20
-    Scenario: Tags 2
-        Given I have a bar
+        @rule_tag_01
+        Rule: Rule tag
+
+            @scenario_tag_10 @scenario_tag_20
+            Scenario: Tags 2
+                Given I have a bar
 
     """,
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         """
         import pytest
         from pytest_bdd import given, scenarios
 
         @given('I have a bar')
-        def i_have_bar():
+        def _():
             return 'bar'
 
         scenarios('test.feature')
     """
     )
-    result = testdir.runpytest("-m", "scenario_tag_10 and not scenario_tag_01", "-vv")
+    result = pytester.runpytest("-m", "scenario_tag_10 and not scenario_tag_01", "-vv")
     outcomes = result.parseoutcomes()
     assert outcomes["passed"] == 1
     assert outcomes["deselected"] == 1
 
-    result = testdir.runpytest("-m", "scenario_tag_01 and not scenario_tag_10", "-vv").parseoutcomes()
+    result = pytester.runpytest("-m", "scenario_tag_01 and not scenario_tag_10", "-vv").parseoutcomes()
     assert result["passed"] == 1
     assert result["deselected"] == 1
 
-    result = testdir.runpytest("-m", "feature_tag_1", "-vv").parseoutcomes()
+    result = pytester.runpytest("-m", "feature_tag_1", "-vv").parseoutcomes()
     assert result["passed"] == 2
 
-    result = testdir.runpytest("-m", "feature_tag_10", "-vv").parseoutcomes()
+    result = pytester.runpytest("-m", "feature_tag_10", "-vv").parseoutcomes()
     assert result["deselected"] == 2
 
+    result = pytester.runpytest("-m", "rule_tag_01", "-vv").parseoutcomes()
+    assert result["deselected"] == 1
 
-def test_tags_after_background_issue_160(testdir):
+
+def test_tags_after_background_issue_160(pytester):
     """Make sure using a tag after background works."""
-    testdir.makefile(
+    pytester.makefile(
         ".ini",
         pytest=textwrap.dedent(
             """
@@ -79,7 +83,7 @@ def test_tags_after_background_issue_160(testdir):
     """
         ),
     )
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         test="""
     Feature: Tags after background
@@ -95,29 +99,29 @@ def test_tags_after_background_issue_160(testdir):
             Given I have a baz
     """,
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         """
         import pytest
         from pytest_bdd import given, scenarios
 
         @given('I have a bar')
-        def i_have_bar():
+        def _():
             return 'bar'
 
         @given('I have a baz')
-        def i_have_baz():
+        def _():
             return 'baz'
 
         scenarios('test.feature')
     """
     )
-    result = testdir.runpytest("-m", "tag", "-vv").parseoutcomes()
+    result = pytester.runpytest("-m", "tag", "-vv").parseoutcomes()
     assert result["passed"] == 1
     assert result["deselected"] == 1
 
 
-def test_apply_tag_hook(testdir):
-    testdir.makeconftest(
+def test_apply_tag_hook(pytester):
+    pytester.makeconftest(
         """
         import pytest
 
@@ -132,7 +136,7 @@ def test_apply_tag_hook(testdir):
                 return None
     """
     )
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         test="""
     Feature: Customizing tag handling
@@ -146,69 +150,24 @@ def test_apply_tag_hook(testdir):
             Given I have a bar
     """,
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         """
         from pytest_bdd import given, scenarios
 
         @given('I have a bar')
-        def i_have_bar():
+        def _():
             return 'bar'
 
         scenarios('test.feature')
     """
     )
-    result = testdir.runpytest("-rsx")
+    result = pytester.runpytest("-rsx")
     result.stdout.fnmatch_lines(["SKIP*: Not implemented yet"])
-    result.stdout.fnmatch_lines(["*= 1 skipped, 1 xpassed * =*"])
+    result.stdout.fnmatch_lines(["*= 1 skipped, 1 xpassed*=*"])
 
 
-def test_tag_with_spaces(testdir):
-    testdir.makefile(
-        ".ini",
-        pytest=textwrap.dedent(
-            """
-    [pytest]
-    markers =
-        test with spaces
-    """
-        ),
-    )
-    testdir.makeconftest(
-        """
-        import pytest
-
-        @pytest.hookimpl(tryfirst=True)
-        def pytest_bdd_apply_tag(tag, function):
-            assert tag == 'test with spaces'
-    """
-    )
-    testdir.makefile(
-        ".feature",
-        test="""
-    Feature: Tag with spaces
-
-        @test with spaces
-        Scenario: Tags
-            Given I have a bar
-    """,
-    )
-    testdir.makepyfile(
-        """
-        from pytest_bdd import given, scenarios
-
-        @given('I have a bar')
-        def i_have_bar():
-            return 'bar'
-
-        scenarios('test.feature')
-    """
-    )
-    result = testdir.runpytest_subprocess()
-    result.stdout.fnmatch_lines(["*= 1 passed * =*"])
-
-
-def test_at_in_scenario(testdir):
-    testdir.makefile(
+def test_at_in_scenario(pytester):
+    pytester.makefile(
         ".feature",
         test="""
     Feature: At sign in a scenario
@@ -220,44 +179,58 @@ def test_at_in_scenario(testdir):
             Given I have a baz
     """,
     )
-    testdir.makepyfile(
+    pytester.makepyfile(
         """
         from pytest_bdd import given, scenarios
 
         @given('I have a foo@bar')
-        def i_have_at():
+        def _():
             return 'foo@bar'
 
         @given('I have a baz')
-        def i_have_baz():
+        def _():
             return 'baz'
 
         scenarios('test.feature')
     """
     )
-
-    # Deprecate --strict after pytest 6.1
-    # https://docs.pytest.org/en/stable/deprecations.html#the-strict-command-line-option
-    pytest_version = pkg_resources.get_distribution("pytest").parsed_version
-    if pytest_version >= pkg_resources.parse_version("6.2"):
-        strict_option = "--strict-markers"
-    else:
-        strict_option = "--strict"
-    result = testdir.runpytest_subprocess(strict_option)
+    strict_option = "--strict-markers"
+    result = pytester.runpytest_subprocess(strict_option)
     result.stdout.fnmatch_lines(["*= 2 passed * =*"])
 
 
-@pytest.mark.parametrize(
-    "line, expected",
-    [
-        ("@foo @bar", {"foo", "bar"}),
-        ("@with spaces @bar", {"with spaces", "bar"}),
-        ("@double @double", {"double"}),
-        ("    @indented", {"indented"}),
-        (None, set()),
-        ("foobar", set()),
-        ("", set()),
-    ],
-)
-def test_get_tags(line, expected):
-    assert get_tags(line) == expected
+def test_multiline_tags(pytester):
+    pytester.makefile(
+        ".feature",
+        test="""
+    Feature: Scenario with tags over multiple lines
+
+        @tag1
+        @tag2
+        Scenario: Tags
+            Given I have a foo
+
+        Scenario: Second
+            Given I have a baz
+    """,
+    )
+    pytester.makepyfile(
+        """
+        from pytest_bdd import given, scenarios
+
+        @given('I have a foo')
+        def _():
+            pass
+
+        @given('I have a baz')
+        def _():
+            pass
+
+        scenarios('test.feature')
+    """
+    )
+    result = pytester.runpytest("-m", "tag1", "-vv")
+    result.assert_outcomes(passed=1, deselected=1)
+
+    result = pytester.runpytest("-m", "tag2", "-vv")
+    result.assert_outcomes(passed=1, deselected=1)
