@@ -1,9 +1,11 @@
 """Code generation and assertion tests."""
+
+from __future__ import annotations
+
 import itertools
 import textwrap
 
 from pytest_bdd.scenario import get_python_name_generator
-from tests.utils import assert_outcomes
 
 
 def test_python_name_generator():
@@ -15,9 +17,9 @@ def test_python_name_generator():
     ]
 
 
-def test_generate_missing(testdir):
+def test_generate_missing(pytester):
     """Test generate missing command."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         generation=textwrap.dedent(
             """\
@@ -40,7 +42,7 @@ def test_generate_missing(testdir):
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import functools
@@ -50,7 +52,7 @@ def test_generate_missing(testdir):
         scenario = functools.partial(scenario, "generation.feature")
 
         @given("I have a bar")
-        def i_have_a_bar():
+        def _():
             return "bar"
 
         @scenario("Scenario tests which are already bound to the tests stay as is")
@@ -64,8 +66,8 @@ def test_generate_missing(testdir):
         )
     )
 
-    result = testdir.runpytest("--generate-missing", "--feature", "generation.feature")
-    assert_outcomes(result, passed=0, failed=0, errors=0)
+    result = pytester.runpytest("--generate-missing", "--feature", "generation.feature")
+    result.assert_outcomes(passed=0, failed=0, errors=0)
     assert not result.stderr.str()
     assert result.ret == 0
 
@@ -80,16 +82,14 @@ def test_generate_missing(testdir):
         ]
     )
 
-    result.stdout.fnmatch_lines(
-        ['Step Given "I have a foobar" is not defined in the background of the feature "Missing code generation" *']
-    )
+    result.stdout.fnmatch_lines(['Background step Given "I have a foobar" is not defined*'])
 
     result.stdout.fnmatch_lines(["Please place the code above to the test file(s):"])
 
 
-def test_generate_missing_with_step_parsers(testdir):
+def test_generate_missing_with_step_parsers(pytester):
     """Test that step parsers are correctly discovered and won't be part of the missing steps."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         generation=textwrap.dedent(
             """\
@@ -104,7 +104,7 @@ def test_generate_missing_with_step_parsers(testdir):
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import functools
@@ -114,30 +114,30 @@ def test_generate_missing_with_step_parsers(testdir):
         scenarios("generation.feature")
 
         @given("I use the string parser without parameter")
-        def i_have_a_bar():
+        def _():
             return None
 
         @given(parsers.parse("I use parsers.parse with parameter {param}"))
-        def i_have_n_baz(param):
+        def _(param):
             return param
 
         @given(parsers.re(r"^I use parsers.re with parameter (?P<param>.*?)$"))
-        def i_have_n_baz(param):
+        def _(param):
             return param
 
         @given(parsers.cfparse("I use parsers.cfparse with parameter {param:d}"))
-        def i_have_n_baz(param):
+        def _(param):
             return param
         """
         )
     )
 
-    result = testdir.runpytest("--generate-missing", "--feature", "generation.feature")
-    assert_outcomes(result, passed=0, failed=0, errors=0)
+    result = pytester.runpytest("--generate-missing", "--feature", "generation.feature")
+    result.assert_outcomes(passed=0, failed=0, errors=0)
     assert not result.stderr.str()
     assert result.ret == 0
 
-    output = result.stdout.str()
+    output = str(result.stdout)
 
     assert "I use the string parser" not in output
     assert "I use parsers.parse" not in output
