@@ -1,5 +1,7 @@
 """Test code generation command."""
 
+from __future__ import annotations
+
 import os
 import sys
 import textwrap
@@ -9,11 +11,11 @@ from pytest_bdd.scripts import main
 PATH = os.path.dirname(__file__)
 
 
-def test_generate(testdir, monkeypatch, capsys):
+def test_generate(pytester, monkeypatch, capsys):
     """Test if the code is generated by a given feature."""
 
-    features = testdir.mkdir("scripts")
-    feature = features.join("generate.feature")
+    features = pytester.mkdir("scripts")
+    feature = features.joinpath("generate.feature")
     feature.write_text(
         textwrap.dedent(
             """\
@@ -29,10 +31,9 @@ def test_generate(testdir, monkeypatch, capsys):
             """
         ),
         "utf-8",
-        ensure=True,
     )
 
-    monkeypatch.setattr(sys, "argv", ["", "generate", feature.strpath])
+    monkeypatch.setattr(sys, "argv", ["", "generate", str(feature)])
     main()
     out, err = capsys.readouterr()
     assert out == textwrap.dedent(
@@ -53,25 +54,25 @@ def test_generate(testdir, monkeypatch, capsys):
 
 
     @given('1 have a fixture (appends 1 to a list) in reuse syntax')
-    def have_a_fixture_appends_1_to_a_list_in_reuse_syntax():
+    def _():
         """1 have a fixture (appends 1 to a list) in reuse syntax."""
         raise NotImplementedError
 
 
     @given('I have an empty list')
-    def i_have_an_empty_list():
+    def _():
         """I have an empty list."""
         raise NotImplementedError
 
 
     @when('I use this fixture')
-    def i_use_this_fixture():
+    def _():
         """I use this fixture."""
         raise NotImplementedError
 
 
     @then('my list should be [1]')
-    def my_list_should_be_1():
+    def _():
         """my list should be [1]."""
         raise NotImplementedError
 
@@ -79,9 +80,9 @@ def test_generate(testdir, monkeypatch, capsys):
     )
 
 
-def test_generate_with_quotes(testdir):
+def test_generate_with_quotes(pytester):
     """Test that code generation escapes quote characters properly."""
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         generate_with_quotes=textwrap.dedent(
             '''\
@@ -100,8 +101,8 @@ def test_generate_with_quotes(testdir):
         ),
     )
 
-    result = testdir.run("pytest-bdd", "generate", "generate_with_quotes.feature")
-    assert result.stdout.str() == textwrap.dedent(
+    result = pytester.run("pytest-bdd", "generate", "generate_with_quotes.feature")
+    assert str(result.stdout) == textwrap.dedent(
         '''\
     """Handling quotes in code generation feature tests."""
 
@@ -119,50 +120,50 @@ def test_generate_with_quotes(testdir):
 
 
     @given('I have a fixture with "double" quotes')
-    def i_have_a_fixture_with_double_quotes():
+    def _():
         """I have a fixture with "double" quotes."""
         raise NotImplementedError
 
 
     @given('I have a fixture with \\'single\\' quotes')
-    def i_have_a_fixture_with_single_quotes():
+    def _():
         """I have a fixture with 'single' quotes."""
         raise NotImplementedError
 
 
     @given('I have a fixture with double-quote """triple""" quotes')
-    def i_have_a_fixture_with_doublequote_triple_quotes():
+    def _():
         """I have a fixture with double-quote \\"\\"\\"triple\\"\\"\\" quotes."""
         raise NotImplementedError
 
 
     @given('I have a fixture with single-quote \\'\\'\\'triple\\'\\'\\' quotes')
-    def i_have_a_fixture_with_singlequote_triple_quotes():
+    def _():
         """I have a fixture with single-quote \'\'\'triple\'\'\' quotes."""
         raise NotImplementedError
 
 
     @when('I generate the code')
-    def i_generate_the_code():
+    def _():
         """I generate the code."""
         raise NotImplementedError
 
 
     @then('The generated string should be written')
-    def the_generated_string_should_be_written():
+    def _():
         """The generated string should be written."""
         raise NotImplementedError
     '''
     )
 
 
-def test_unicode_characters(testdir, monkeypatch):
+def test_unicode_characters(pytester, monkeypatch):
     """Test generating code with unicode characters.
 
     Primary purpose is to ensure compatibility with Python2.
     """
 
-    testdir.makefile(
+    pytester.makefile(
         ".feature",
         unicode_characters=textwrap.dedent(
             """\
@@ -176,10 +177,7 @@ def test_unicode_characters(testdir, monkeypatch):
         ),
     )
 
-    if sys.version_info < (3, 7):
-        monkeypatch.setenv("PYTHONIOENCODING", "utf-8")
-
-    result = testdir.run("pytest-bdd", "generate", "unicode_characters.feature")
+    result = pytester.run("pytest-bdd", "generate", "unicode_characters.feature")
     expected_output = textwrap.dedent(
         '''\
             """Generating unicode characters feature tests."""
@@ -198,21 +196,21 @@ def test_unicode_characters(testdir, monkeypatch):
 
 
             @given('We have a circle')
-            def we_have_a_circle():
+            def _():
                 """We have a circle."""
                 raise NotImplementedError
 
 
             @when('We want to know its circumference')
-            def we_want_to_know_its_circumference():
+            def _():
                 """We want to know its circumference."""
                 raise NotImplementedError
 
 
             @then('We calculate 2 * ℼ * 𝑟')
-            def we_calculate_2__ℼ__𝑟():
+            def _():
                 """We calculate 2 * ℼ * 𝑟."""
                 raise NotImplementedError
             '''
     )
-    assert result.stdout.str() == expected_output
+    assert str(result.stdout) == expected_output
