@@ -1,10 +1,12 @@
 """Tests for testing cases when we have unicode in feature file."""
 
+from __future__ import annotations
+
 import textwrap
 
 
-def test_steps_in_feature_file_have_unicode(testdir):
-    testdir.makefile(
+def test_steps_in_feature_file_have_unicode(pytester):
+    pytester.makefile(
         ".feature",
         unicode=textwrap.dedent(
             """\
@@ -21,7 +23,7 @@ def test_steps_in_feature_file_have_unicode(testdir):
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import sys
@@ -38,7 +40,7 @@ def test_steps_in_feature_file_have_unicode(testdir):
 
 
         @given(parsers.parse(u"у мене є рядок який містить '{content}'"))
-        def there_is_a_string_with_content(content, string):
+        def _(content, string):
             string["content"] = content
 
 
@@ -46,17 +48,17 @@ def test_steps_in_feature_file_have_unicode(testdir):
 
 
         @then(parsers.parse("I should see that the string equals to content '{content}'"))
-        def assert_that_the_string_equals_to_content(content, string):
+        def _(content, string):
             assert string["content"] == content
         """
         )
     )
-    result = testdir.runpytest()
+    result = pytester.runpytest()
     result.assert_outcomes(passed=1)
 
 
-def test_steps_in_py_file_have_unicode(testdir):
-    testdir.makefile(
+def test_steps_in_py_file_have_unicode(pytester):
+    pytester.makefile(
         ".feature",
         unicode=textwrap.dedent(
             """\
@@ -69,7 +71,7 @@ def test_steps_in_py_file_have_unicode(testdir):
         ),
     )
 
-    testdir.makepyfile(
+    pytester.makepyfile(
         textwrap.dedent(
             """\
         import pytest
@@ -85,15 +87,15 @@ def test_steps_in_py_file_have_unicode(testdir):
 
 
         @given("there is an other string with content 'якийсь контент'")
-        def there_is_an_other_string_with_content(string):
+        def _(string):
             string["content"] = u"с каким-то контентом"
 
         @then("I should see that the other string equals to content 'якийсь контент'")
-        def assert_that_the_other_string_equals_to_content(string):
+        def _(string):
             assert string["content"] == u"с каким-то контентом"
 
         """
         )
     )
-    result = testdir.runpytest()
+    result = pytester.runpytest()
     result.assert_outcomes(passed=1)
